fix: reorganize component exports in client.ts and index.ts
- Added new exports for various Radix UI components in client.ts, enhancing the component library.
- Removed outdated exports from index.ts to streamline the component structure and improve clarity.
- Introduced Card component exports to replace removed components, ensuring continuity in the library's offerings.
