+++ src/components/client.ts
@@ -14,3 +14,125 @@ export {
   AlertDialogAction,
   AlertDialogCancel,
 } from './AlertDialog/AlertDialog';
+
+export { Calendar, CalendarDayButton } from './Calendar/Calendar';
+export {
+  type CarouselApi,
+  Carousel,
+  CarouselContent,
+  CarouselItem,
+  CarouselPrevious,
+  CarouselNext,
+} from './Carousel/Carousel';
+export { Combobox, type ComboboxOption } from './Combobox/Combobox';
+export { InputOTP, InputOTPGroup, InputOTPSlot, InputOTPSeparator } from './InputOTP/InputOTP';
+export { Slider } from './Slider/Slider';
+export { Progress } from './Progress/Progress';
+
+export { default as LanguageSelector } from './LanguageSelector/LanguageSelector';
+export { Marquee } from './Marquee/Marquee';
+
+// Radix UI components that require client-side functionality
+export {
+  Accordion,
+  AccordionItem,
+  AccordionTrigger,
+  AccordionContent,
+} from './Accordion/Accordion';
+
+export { Avatar, AvatarImage, AvatarFallback } from './Avatar/Avatar';
+export { Checkbox } from './Checkbox/Checkbox';
+
+export {
+  ContextMenu,
+  ContextMenuTrigger,
+  ContextMenuContent,
+  ContextMenuItem,
+  ContextMenuCheckboxItem,
+  ContextMenuRadioItem,
+  ContextMenuLabel,
+  ContextMenuSeparator,
+  ContextMenuShortcut,
+  ContextMenuGroup,
+  ContextMenuPortal,
+  ContextMenuSub,
+  ContextMenuSubContent,
+  ContextMenuSubTrigger,
+  ContextMenuRadioGroup,
+} from './ContextMenu/ContextMenu';
+
+export {
+  DropdownMenu,
+  DropdownMenuPortal,
+  DropdownMenuTrigger,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuLabel,
+  DropdownMenuItem,
+  DropdownMenuCheckboxItem,
+  DropdownMenuRadioGroup,
+  DropdownMenuRadioItem,
+  DropdownMenuSeparator,
+  DropdownMenuShortcut,
+  DropdownMenuSub,
+  DropdownMenuSubTrigger,
+  DropdownMenuSubContent,
+} from './DropdownMenu/DropdownMenu';
+
+export { HoverCard, HoverCardTrigger, HoverCardContent } from './HoverCard/HoverCard';
+export { RadioGroup, RadioGroupItem } from './RadioGroup/RadioGroup';
+export { ScrollArea, ScrollBar } from './ScrollArea/ScrollArea';
+export { Switch } from './Switch/Switch';
+export { Tabs, TabsList, TabsTrigger, TabsContent } from './Tabs/Tabs';
+export { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from './Tooltip/Tooltip';
+
+export {
+  Command,
+  CommandDialog,
+  CommandInput,
+  CommandList,
+  CommandEmpty,
+  CommandGroup,
+  CommandItem,
+  CommandShortcut,
+  CommandSeparator,
+} from './Command/Command';
+
+export {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+} from './Dialog/Dialog';
+
+export {
+  Drawer,
+  DrawerPortal,
+  DrawerOverlay,
+  DrawerTrigger,
+  DrawerClose,
+  DrawerContent,
+  DrawerHeader,
+  DrawerFooter,
+  DrawerTitle,
+  DrawerDescription,
+} from './Drawer/Drawer';
+
+export {
+  Sheet,
+  SheetTrigger,
+  SheetClose,
+  SheetContent,
+  SheetHeader,
+  SheetFooter,
+  SheetTitle,
+  SheetDescription,
+} from './Sheet/Sheet';
+
+export { Toaster } from './Toaster/Toaster';
