+++ src/components/index.ts
@@ -54,14 +54,7 @@ export { default as FormattedDate } from './FormattedDate/FormattedDate';
 export { default as FormattedNumber } from './FormattedNumber/FormattedNumber';
 export { default as LanguageSelector } from './LanguageSelector/LanguageSelector';
 export { Marquee } from './Marquee/Marquee';
-export {
-  Accordion,
-  AccordionItem,
-  AccordionTrigger,
-  AccordionContent,
-} from './Accordion/Accordion';
 export { Alert, AlertTitle, AlertDescription } from './Alert/Alert';
-export { Avatar, AvatarImage, AvatarFallback } from './Avatar/Avatar';
 export { Badge } from './Badge/Badge';
 export {
   Breadcrumb,
@@ -72,88 +65,15 @@ export {
   BreadcrumbSeparator,
   BreadcrumbEllipsis,
 } from './Breadcrumb/Breadcrumb';
-export { Calendar, CalendarDayButton } from './Calendar/Calendar';
 export {
-  type CarouselApi,
-  Carousel,
-  CarouselContent,
-  CarouselItem,
-  CarouselPrevious,
-  CarouselNext,
-} from './Carousel/Carousel';
-export { Checkbox } from './Checkbox/Checkbox';
-export { Combobox, type ComboboxOption } from './Combobox/Combobox';
-export {
-  Command,
-  CommandDialog,
-  CommandInput,
-  CommandList,
-  CommandEmpty,
-  CommandGroup,
-  CommandItem,
-  CommandShortcut,
-  CommandSeparator,
-} from './Command/Command';
-export {
-  ContextMenu,
-  ContextMenuTrigger,
-  ContextMenuContent,
-  ContextMenuItem,
-  ContextMenuCheckboxItem,
-  ContextMenuRadioItem,
-  ContextMenuLabel,
-  ContextMenuSeparator,
-  ContextMenuShortcut,
-  ContextMenuGroup,
-  ContextMenuPortal,
-  ContextMenuSub,
-  ContextMenuSubContent,
-  ContextMenuSubTrigger,
-  ContextMenuRadioGroup,
-} from './ContextMenu/ContextMenu';
-export {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogOverlay,
-  DialogPortal,
-  DialogTitle,
-  DialogTrigger,
-} from './Dialog/Dialog';
-export {
-  Drawer,
-  DrawerPortal,
-  DrawerOverlay,
-  DrawerTrigger,
-  DrawerClose,
-  DrawerContent,
-  DrawerHeader,
-  DrawerFooter,
-  DrawerTitle,
-  DrawerDescription,
-} from './Drawer/Drawer';
-export {
-  DropdownMenu,
-  DropdownMenuPortal,
-  DropdownMenuTrigger,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuLabel,
-  DropdownMenuItem,
-  DropdownMenuCheckboxItem,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuShortcut,
-  DropdownMenuSub,
-  DropdownMenuSubTrigger,
-  DropdownMenuSubContent,
-} from './DropdownMenu/DropdownMenu';
-export { HoverCard, HoverCardTrigger, HoverCardContent } from './HoverCard/HoverCard';
-export { InputOTP, InputOTPGroup, InputOTPSlot, InputOTPSeparator } from './InputOTP/InputOTP';
+  Card,
+  CardContent,
+  CardHeader,
+  CardFooter,
+  CardTitle,
+  CardDescription,
+  CardAction,
+} from './Card/Card';
 export { Label } from './Label/Label';
 export {
   Pagination,
@@ -164,24 +84,8 @@ export {
   PaginationNext,
   PaginationEllipsis,
 } from './Pagination/Pagination';
-export { Progress } from './Progress/Progress';
-export { RadioGroup, RadioGroupItem } from './RadioGroup/RadioGroup';
-export { ScrollArea, ScrollBar } from './ScrollArea/ScrollArea';
 export { Separator } from './Separator/Separator';
-export {
-  Sheet,
-  SheetTrigger,
-  SheetClose,
-  SheetContent,
-  SheetHeader,
-  SheetFooter,
-  SheetTitle,
-  SheetDescription,
-} from './Sheet/Sheet';
 export { Skeleton } from './Skeleton/Skeleton';
-export { Slider } from './Slider/Slider';
-export { Toaster } from './Toaster/Toaster';
-export { Switch } from './Switch/Switch';
 export {
   Table,
   TableHeader,
@@ -192,7 +96,4 @@ export {
   TableCell,
   TableCaption,
 } from './Table/Table';
-export { Tabs, TabsList, TabsTrigger, TabsContent } from './Tabs/Tabs';
 export { Textarea } from './Textarea/Textarea';
-export { Toggle } from './Toggle/Toggle';
-export { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from './Tooltip/Tooltip';
